test: DAO 변경에 따른 테스트 코드 변경
